requirements.txt
# Core ML
torch>=2.1.0
safetensors>=0.4.0
sentencepiece>=0.1.99
transformers>=4.36.0
bitsandbytes>=0.41.0
accelerate>=0.25.0

# Vision
open-clip-torch>=2.24.0
Pillow>=10.0.0

# API
fastapi[standard]>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Memory / Vector Store
chromadb>=0.4.0

# Utilities
numpy>=1.24.0
scipy>=1.11.0
aiofiles>=23.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
